rexstan + ymca best practices

+++ update.php
@@ -1,7 +1,9 @@
 <?php
 
 /* Tablesets aktualisieren */
+$addon = rex_addon::get('yrewrite_metainfo');
 
-rex_yform_manager_table_api::importTablesets(rex_file::get(rex_path::addon($this->name, 'install/rex_yrewrite_domain_meta.tableset.json')));
-
-rex_yform_manager_table::deleteCache();
+if (rex_addon::get('yform')->isAvailable() && !rex::isSafeMode()) {
+    rex_yform_manager_table_api::importTablesets(rex_file::get(__DIR__ . '/install/rex_yrewrite_domain_meta.tableset.json')); // @phpstan-ignore-line
+    rex_yform_manager_table::deleteCache();
+}
